S. Is Triangle
Given three numbers A, B and C that donate the lengths of the three sides of a triangle. Determine whether this triangle is valid or not and if it is valid print its area.

Input
Only one line contains three numbers A, B and C (1 ≤ A, B, C ≤ 103).

Output
If the three sides preform a triangle print "Valid" and the area of the triangle in separate lines otherwise, print "Invalid".

Don't print any extra spaces.